Fix broken view selection API
In the prevous version, AlgorithmParameters.viewSelection was set to
true by default, and the user was provided with a flag that
(redundantly) enables it, but there was no way to disable it.

There is probably a better fix, but IMHO argument parsing needs to be
overhauled anyway; there are libraries for sane argparsing.